backend/requirements.txt
absl-py==2.4.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
astunparse==1.6.3
attrs==25.4.0
Authlib==1.6.6
bcrypt==4.1.3
beautifulsoup4==4.14.3
black==25.12.0
blinker==1.9.0
boto3==1.42.29
botocore==1.42.29
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cmake==4.2.1
cryptography==46.0.3
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
emergentintegrations==0.1.0
et_xmlfile==2.0.0
fastapi==0.110.1
fastuuid==0.14.0
filelock==3.20.3
fire==0.7.1
flake8==7.3.0
Flask==3.1.2
flask-cors==6.0.2
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.1.0
gast==0.7.0
gdown==5.2.1
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
google-genai==1.59.0
google-generativeai==0.8.6
google-pasta==0.2.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
gunicorn==25.0.3
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.1
httpx==0.28.1
huggingface_hub==1.3.2
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
isort==7.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
joblib==1.5.3
jq==1.10.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
libclang==18.1.1
librt==0.7.8
lightdsa==0.0.3
lightecc==0.0.4
lightphe==0.0.20
litellm==1.80.0
lxml==6.0.2
lz4==4.4.5
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
motor==3.3.1
mpmath==1.3.0
multidict==6.7.0
mypy==1.19.1
mypy_extensions==1.1.0
namex==0.1.0
numpy==2.4.1
oauthlib==3.3.1
openai==1.99.9
opencv-python-headless==4.13.0.90
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pathspec==1.0.3
pdf2image==1.17.0
pillow==12.1.0
platformdirs==4.5.1
pluggy==1.6.0
propcache==0.4.1
proto-plus==1.27.0
protobuf==5.29.5
psutil==7.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.5.0
pyotp==2.9.0
pyparsing==3.3.1
PyPDF2==3.0.1
PySocks==1.7.1
pytest==9.0.2
python-barcode==0.16.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.21
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
qrcode==8.2
referencing==0.37.0
regex==2026.1.15
reportlab==4.4.9
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.16.0
s5cmd==0.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
starlette==0.37.2
stripe==14.1.0
sympy==1.14.0
tenacity==9.1.2
termcolor==3.3.0
tiktoken==0.12.0
tokenizers==0.22.2
tqdm==4.67.1
typer==0.21.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.25.0
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==3.1.5
wrapt==2.1.1
xlsxwriter==3.2.9
yarl==1.22.0
zipp==3.23.0
